# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Disbursement API",
        "contact": {
            "name": "Disbursement Team",
            "email": "support@email.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/seller/register": {
            "post": {
                "tags": [
                    "seller"
                ],
                "operationId": "/seller/register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "seller_name": {
                                        "type": "string"
                                    },
                                    "seller_email": {
                                        "type": "string"
                                    },
                                    "seller_bank_code": {
                                        "type": "string"
                                    },
                                    "seller_account_name": {
                                        "type": "string"
                                    },
                                    "seller_account_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns succes register seller deposit balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/registerseller"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/emptyarray"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "emptyarray": {},
...
            "registerseller": {
                "properties": {
                    "seller_name": {
                        "type": "string"
                    },
                    "seller_bank_code": {
                        "type": "string"
                    },
                    "seller_account_name": {
                        "type": "string"
                    },
                    "seller_account_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
...
    }
}